worker/apiary-worker/pyproject.toml
[tool.poetry]
name = "apiary-worker"
version = "0.1.0"
description = ""
authors = ["PiloeGAO <leonumerique@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
websocket-client = "^1.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
